+++ OwnTube.tv/lib/videoServices.ts
@@ -1,0 +1,38 @@
+import * as testData from "../testData.json";
+import { Video, Category } from "../types";
+
+class VideoService {
+  private videos: Video[] = [];
+  private categories: Category[] = [];
+  private readonly baseThumbnailUrl = "https://peertube2.cpy.re";
+
+  constructor() {
+    this.loadVideosFromJson();
+  }
+
+  public getVideoCategoryLabels(): string[] {
+    return this.categories.map(({ id, label }) => label);
+  }
+
+  public completeThumbnailUrls(): Video[] {
+    return this.videos.map((video) => ({
+      ...video,
+      thumbnailUrl: `${this.baseThumbnailUrl}${video.thumbnailPath}`,
+    }));
+  }
+
+  public getVideosForCategory(categoryLabel: string): Video[] {
+    return this.videos.filter((video) => video.category.label === categoryLabel);
+  }
+
+  public loadVideosFromJson(): void {
+    const data = testData;
+    this.videos = data.data;
+    const uniqueCategories: Category[] = Array.from(new Set(this.videos.map((video) => video.category.label))).map(
+      (label, index) => ({ id: index + 1, label }),
+    );
+    this.categories = uniqueCategories;
+  }
+}
+
+export default VideoService;
